mysite/shopapp/serializers.py
from .models import Product, Order
from rest_framework import serializers


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = (
            'pk',
            'name',
            'price',
            'description',
            'created_at',
            'discount',
            'archived',
            'preview'
        )


class OrderSerializer(serializers.ModelSerializer):
    class Meta:
        model = Order
        fields = (
            'pk',
            'delivery_address',
            'user',
            'promocode',
            'created_at',
            'receipt'
        )

