requirements.txt
asgiref==3.5.2
Django==4.1.7
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.2
django-filter==23.2
djangorestframework==3.14.0

